+++ CHANGELOG.md
@@ -60,6 +60,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - [Defaults] Test and document `ValueError`s that can be raised by functions in `iati.default`. [#241]
 
 - [Rulesets] `name` attribute on a Rule changed to read-only property. [#251]
+- [Rulesets] Equal Rulesets are now deemed to be equal. [#249]
 
 - [Validation] Prevent `XPathEvalError`s occurring when given a Codelist Mapping XPath that identifies something other than an attribute. [#229]
 - [Validation] Datasets with an `xml:lang` attribute no longer raise a `KeyError` upon performing Codelist validation against a Schema populated with the Language Codelist. [#226]